makefile commit

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: jcharnec <jcharnec@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/08/17 15:23:07 by jcharnec          #+#    #+#             */
-/*   Updated: 2022/09/20 11:56:15 by jcharnec         ###   ########.fr       */
+/*   Updated: 2022/09/20 12:17:45 by jcharnec         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
